move fezing to new function and restore the other
